Create README.md
In this project I have done an attendance log application, it can be used in any conference, career fair, supercenter, classroom etc. This program will store personal detail to keep track about attendance (name, Id or phone no and email). We can store information and then can be searched by a person's name to see the details later time as needed. Also, sort by ascending order and as well delete any name from the list too.
It will be updated over time. I am still thinking about the date what we can capture to search an find out details by date and time. Also, it can be store as text format to the local drive. For this class project, we get the idea of real-world settings and experience. Where how to work and deal with that's helpful as like real experience. After running this program, we can see the display option where to have to add, search, delete and sort button. When we push the add button it will pop up an input dialog box to store data and will save it. When we push the search option, it will pop up a box to type the name of the search and after type when press ok the search details will appear on the top. Same way to delete any name from the list. To sort need to push the sort button.
A very simple program, but it can be used in various ways. Can be daily use in restricted areas to store visitor detail every day and big shopping mall for workers and vendors who are visiting or in the building. Will be working on to do better may need to change some to fit better for the next update version.